Create websocket.test.js
In this test, a WebSocket server is created alongside an HTTP server to listen for incoming connections. Then, two WebSocket clients are created (clientA and clientB) that join different rooms. We test that when clientA sends a message to its room, it receives the message, and clientB does not receive the message since it's in a different room.